Replace int() by float() in MakeWlmFile1.py to avoid the error when convert a string to int when the string represents a floating point number

+++ Utilities/MakeWlmFile1.py
@@ -85,7 +85,7 @@ class WlmFileMaker(object):
         self.fp = file(self.fname,"w")
 
         if "-w" in options:
-            self.waitTime = int(options["-w"])
+            self.waitTime = float(options["-w"])
         else:
             self.waitTime = float(self.config["SETTINGS"]["WAIT_TIME"])
         if self.waitTime < 0:
